@@ -2,9 +2,9 @@ import 'hits.dart';
 
 class PixabayResult {
   PixabayResult({
-    int total,
-    int totalHits,
-    List<Hits> hits,
+    int? total,
+    int? totalHits,
+    List<Hits>? hits,
   }) {
     _total = total;
     _totalHits = totalHits;
@@ -17,27 +17,27 @@ class PixabayResult {
     if (json['hits'] != null) {
       _hits = [];
       json['hits'].forEach((v) {
-        _hits.add(Hits.fromJson(v));
+        _hits!.add(Hits.fromJson(v));
       });
     }
   }
 
-  int _total;
-  int _totalHits;
-  List<Hits> _hits;
+  int? _total;
+  int? _totalHits;
+  List<Hits>? _hits;
 
-  int get total => _total;
+  int? get total => _total;
 
-  int get totalHits => _totalHits;
+  int? get totalHits => _totalHits;
 
-  List<Hits> get hits => _hits;
+  List<Hits>? get hits => _hits;
 
   Map<String, dynamic> toJson() {
     final map = <String, dynamic>{};
     map['total'] = _total;
     map['totalHits'] = _totalHits;
     if (_hits != null) {
-      map['hits'] = _hits.map((v) => v.toJson()).toList();
+      map['hits'] = _hits!.map((v) => v.toJson()).toList();
     }
     return map;
   }
