@@ -4,7 +4,7 @@ import 'package:image_searching_app/model/hits.dart';
 class ImageItem extends StatelessWidget {
   final Hits hits;
 
-  const ImageItem({Key key, @required this.hits}) : super(key: key);
+  const ImageItem({Key? key, required this.hits}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -23,13 +23,13 @@ class ImageItem extends StatelessWidget {
               child: Column(
                 children: [
                   Image.network(
-                    hits.previewURL,
+                    hits.previewURL!,
                     width: 200,
                     height: 160,
                     fit: BoxFit.cover,
                   ),
                   SizedBox(height: 5,),
-                  Text(hits.tags,
+                  Text(hits.tags!,
                   style: TextStyle(
                     fontSize: 17
                   ),),
